requirements.txt
scrapy==2.11.2
pymongo==4.8.0
python-dotenv==1.0.1
